apps/workbench-api/requirements.txt
# ============================================================================
# Data Engineering Workbench - Python Dependencies
# ============================================================================

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Redis
redis>=5.0.0
aioredis>=2.0.1

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
duckdb>=0.9.0
pyarrow>=14.0.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Odoo Integration
OdooRPC>=0.9.0

# MinIO / S3
minio>=7.2.0

# Temporal
temporalio>=1.4.0

# Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Security
cryptography>=41.0.0

# Logging & Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0

# Jupyter Integration
jupyter-client>=8.6.0
nbformat>=5.9.0

# Utils
python-dotenv>=1.0.0
pyyaml>=6.0.1
orjson>=3.9.0
tenacity>=8.2.0
python-dateutil>=2.8.2

# Development (comment out in production)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
# black>=23.11.0
# ruff>=0.1.6
# mypy>=1.7.0
